requirements.txt
pep8>=1.3
pytest
django >= 1.6
djangorestframework
django-shortuuidfield
python-openid
django-openid-auth
django-notification
